requirements.txt
accelerate==0.20.3
albumentations==1.3.1
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docker-pycreds==0.4.0
filelock==3.12.2
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.16.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
peft==0.3.0
Pillow==10.0.0
protobuf==4.23.3
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.27.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.7.4
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wandb==0.15.4
